chore: move enclave print functions to separate file
